moving utils to constants
