+++ packages/webapp/src/components/Map/DrawingManager/index.js
@@ -4,7 +4,6 @@ import styles from './styles.module.scss';
 import { useTranslation } from 'react-i18next';
 import { ReactComponent as BackIcon } from '../../../assets/images/map/back.svg';
 import Button from '../../Form/Button';
-import { isArea, isLine, isPoint } from '../../../containers/Map/util';
 
 export default function PureDrawingManager({
   className,
